update DashboardWallet Component with SearchValue props
