+++ src/index.css
@@ -46,6 +46,7 @@ input[type="text"] {
     border: none;
     border-radius: 4px;
     margin-right: 10px;
+    width: 350px;
 }
 
 button {
@@ -61,6 +62,32 @@ button:hover {
     background-color: #005fa3;
 }
 
+.header-first {
+  display: flex;
+  align-items: center;
+  height: 60px;
+  width: 100%;
+}
+
+.nav-first {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+}
+
+.nav-first a {
+  text-decoration: none;
+  color: white;
+  font-size: 1rem;
+  padding: 10px 15px;
+  border-radius: 5px;
+}
+.nav-left, .nav-right {
+  display: flex;
+  align-items: center;
+  gap: 25px;
+}
 /* end nav */
 
 /* second nav */
